requirements.txt
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0 